Filter loci and samples (requires bcftools and tabix to be installed)

This function filters the SNPs and samples based on the specified conditions

Usage

filter_snps_samples(
  snpdata,
  min_qual = 10L,
  max_missing_sites = 0.2,
  max_missing_samples = 0.2,
  maf_cutoff = 0.01
)

Arguments

snpdata

a SNPdata object

min_qual

the minimum call quality score below which a loci will be discarded. default = 10

max_missing_sites

the maximum fraction of missing sites above which a sample should be discarded. default = 0.2

max_missing_samples

the maximum fraction of missing samples above which a loci should be discarded. default = 0.2

maf_cutoff

the MAF cut-off. loci with a MAF < maf_cutoff will be discarded

Value

a filtered SNPdata object

Examples

if (FALSE) {
 snpdata <- filter_snps_samples(
  snpdata,
  min_qual            = 10,
  max_missing_sites   = 0.2,
  max_missing_samples = 0.2,
  maf_cutoff          = 0.01
 )
 }