Create the SNPdata object

This function generate the input data needed for whole genome SNPs data genotyped from malaria parasite.

get_snpdata(
  vcf_file = NULL,
  meta_file = NULL,
  output_dir = NULL,
  gof = NULL,
  gff = NULL,
  num_threads = 4L
)

Arguments

vcf_file

the input VCF file (required)

meta_file

the metadata file (required)

output_dir

the path to the folder where the output files will be stored (optional)

gof

the gene ontology annotation file (optional). If not provided, the default file obtained from the PlasmoDB # nolint: line_length_linter will be used

gff

the gene annotation file (optional). If not provided, the default file obtained from the PlasmoDB # nolint: line_length_linter will be used

num_threads

the number of threads to be used when reading in the data from the VCF file. default is 4

Value

an object of class SNPdata with 5 elements

meta: a data.frame that contains the sample's metadata

details: a data.frame with SNPs genomic coordinates, the fraction of missing data per SNP and the names and descriptions of the gene on which they are located.

GT: an integer matrix with the genotype data. 0='reference allele', 1='alternate allele', 2='mixed allele', NA='missing allele'

vcf: the full path to the VCF file from which the data is generated

index: an integer. this is used as some kind of label that can tell the how many times the VCF files has been modified.

Examples

if (FALSE) {
  snpdata <- get_snpdata(
    vcf_file   = system.file("extdata", "Input_Data.vcf.gz",
                             package = "mpbr"),
    meta_file  = system.file("extdata", "SampleMetadata.RDS",
                             package = "mpbr"),
    output_dir = tempdir()
 )
}